refactored client and host reservation page bindings
